+++ server/CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Agent Server
+
+### Bugfixes
+
+- Changed the Resource instantiation for OTEL based telemetry. ([GPT-1027](https://linear.app/sema4ai/issue/GPT-1027))
+
 # Sema4.ai Agent Server 2.0.5-alpha.1 (2025-06-24)
 
 ## Agent Server